Add modeless background task monitor dialog
