+++ DepanFxApp/src/main/java/com/pnambic/depanfx/app/DepanFxConfiguration.java
@@ -4,6 +4,8 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+import java.time.Clock;
+
 import net.rgielen.fxweaver.core.FxWeaver;
 
 @Configuration
@@ -12,5 +14,10 @@ public class DepanFxConfiguration {
   @Bean
   public FxWeaver fxWeaver(ConfigurableApplicationContext applicationContext) {
     return new FxWeaver(applicationContext::getBean, applicationContext::close);
+  }
+
+  @Bean
+  public Clock depanClock() {
+    return Clock.systemDefaultZone();
   }
 }
