+++ DepanFxTasksGui/src/main/java/com/pnambic/depanfx/tasks/gui/TaskMonitorDialogService.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2025 The Depan Project Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.pnambic.depanfx.tasks.gui;
 
 import com.pnambic.depanfx.scene.DepanFxDialogRunner;
@@ -63,8 +78,8 @@ public class TaskMonitorDialogService {
       return dialogStage;
     }
 
-    DepanFxDialogRunner.Dialog<TaskMonitorDialogController> dialog =
-        dialogRunner.createDialogAndParent(TaskMonitorDialogController.class);
+    DepanFxDialogRunner.Dialog<TaskMonitorDialog> dialog =
+        dialogRunner.createDialogAndParent(TaskMonitorDialog.class);
     Stage stage = dialog.runModeless("Background Tasks");
     stage.setOnHidden(event -> dialogStage = null);
     stage.setOnCloseRequest(event -> dialogStage = null);
